requirements.txt
pyyaml
pandas
pydantic
elasticsearch